Actually add other items to the stack.
Previously this wasn't occuring, so any items that weren't merged
(either by being added into the previous operator immediately above
this, or by having been rearranged as part of a the creation of a
new Operator inside _addOperator) were being lost.

This was then being masked by the (now fixed) bug where Values and
Symbols were claiming to be equal to None.